+++ releaf-core/app/builders/releaf/builders/orderer.rb
@@ -1,5 +1,0 @@
-module Releaf::Builders::Orderer
-  def orderer(items)
-    Releaf::ItemOrderer.new(*items.to_a)
-  end
-end