+++ releaf-core/app/builders/releaf/builders/table_builder.rb
@@ -1,7 +1,6 @@
 class Releaf::Builders::TableBuilder
   include Releaf::Builders::Base
   include Releaf::Builders::Toolbox
-  include Releaf::Builders::Orderer
   attr_accessor :collection, :options, :template, :resource_class
 
   def initialize(collection, resource_class, template, options)