+++ releaf-core/app/lib/releaf/responders/after_save_responder.rb
@@ -12,7 +12,7 @@ module Releaf::Responders
       elsif options[:redirect]
         render json: {url: resource_location}, status: 303
       else
-        redirect_to resource_location, status: 303
+        redirect_to resource_location, status: 303, turbolinks: false
       end
     end