+++ releaf-core/spec/services/array/reorder_spec.rb
@@ -1,57 +1,50 @@
 require "rails_helper"
 
-describe Releaf::ItemOrderer do
-  let(:list){ ["a", "b", "c", "d", {items: "xx"}, "e", "f", "g"] }
-  subject{ described_class.new(*list) }
+describe Array::Reorder do
+  subject{ described_class.new(array: ["a", "b", "c", "d", {items: "xx"}, "e", "f", "g"],
+                               values: "c", options: :first) }
 
-  describe "#initialize" do
-    it "assigns given arguments to list accessor" do
-      subject = described_class.new(:a, :b, :c)
-      expect(subject.list).to eq([:a, :b, :c])
+  describe "#values=" do
+    context "when `values` is not array" do
+      it "wraps it within array before assigning" do
+        expect{ subject.values = :a }.to change{ subject.values }.to([:a])
+      end
+    end
+
+    context "when `values` is array" do
+      it "does not modify it before assigning" do
+        expect{ subject.values = [:a] }.to change{ subject.values }.to([:a])
+      end
     end
   end
 
-  describe "#result" do
-    it "returns list" do
-      expect(subject.result).to eq(list)
-      subject.reorder(:a, :last)
-      expect(subject.result).to eq(["b", "c", "d", {items: "xx"}, "e", "f", "g", "a"])
-    end
-  end
-
-  describe "#to_a" do
-    it "returns result" do
-      allow(subject).to receive(:result).and_return(:x)
-      expect(subject.to_a).to eq(:x)
-    end
-  end
-
-  describe "#reorder" do
-    it "returns instance of itself" do
-      expect(subject.reorder(:a, :last)).to eq(subject)
-    end
-
+  describe "#call" do
     it "deletes given values and insert deleted values in reversed order at reorder index" do
       deleted_values = {a: "x", b: "y"}
+      subject.values = [:a, :b]
+      subject.options = :last
 
       expect(subject).to receive(:delete).with([:a, :b]).ordered.and_return(deleted_values)
       expect(subject).to receive(:reorder_index).with(:last).ordered.and_return(1)
-
-      expect{ subject.reorder([:a, :b], :last) }.to change{ subject.list }
+      expect{ subject.call }.to change{ subject.array }
         .to(["a", "x", "y", "b", "c", "d", {items: "xx"}, "e", "f", "g"])
     end
 
     context "when given value is array" do
       it "process unmodified given value" do
+        subject.values = [:a]
+        subject.options = :last
         expect(subject).to receive(:delete).with([:a]).and_call_original
-        subject.reorder([:a], :last)
+        subject.call
       end
     end
 
     context "when given value is not array" do
       it "puts value within array" do
+        subject.values = [:a]
+        subject.options = :last
         expect(subject).to receive(:delete).with([:a]).and_call_original
-        subject.reorder(:a, :last)
+        subject.call
       end
     end
   end
@@ -64,9 +57,9 @@ describe Releaf::ItemOrderer do
     end
 
     context "when given options is :last" do
-      it "returns size of current list array" do
+      it "returns size of array" do
         expect(subject.reorder_index(:last)).to eq(8)
-        subject.list = [:a, :b]
+        subject.array = [:a, :b]
         expect(subject.reorder_index(:last)).to eq(2)
       end
     end
@@ -93,19 +86,19 @@ describe Releaf::ItemOrderer do
   end
 
   describe "#index" do
-    it "returns value index by comparing list and given values casted to strings" do
+    it "returns value index by comparing array and given values casted to strings" do
       expect(subject.index(:c)).to eq(2)
       expect(subject.index("c")).to eq(2)
-      subject.list = ["a", :b, :c]
+      subject.array = ["a", :b, :c]
       expect(subject.index("c")).to eq(2)
     end
 
-    context "when list value is Hash" do
+    context "when array value is Hash" do
       it "compares by hash first key casted to string" do
         expect(subject.index(:items)).to eq(4)
         expect(subject.index("items")).to eq(4)
 
-        subject.list = ["a", {"items" => "x"}]
+        subject.array = ["a", {"items" => "x"}]
         expect(subject.index(:items)).to eq(1)
         expect(subject.index("items")).to eq(1)
       end
@@ -114,29 +107,15 @@ describe Releaf::ItemOrderer do
 
   describe "#delete" do
     before do
-      subject.list = [:a, :b, :c, :d, {items: "x"}]
+      subject.array = [:a, :b, :c, :d, {items: "x"}]
     end
 
-    it "deletes given values from list" do
-      expect{ subject.delete(["a", "b", "items"]) }.to change{ subject.list }.to([:c, :d])
+    it "deletes given values from array" do
+      expect{ subject.delete(["a", "b", "items"]) }.to change{ subject.array }.to([:c, :d])
     end
 
     it "returns hash with mapped deleted values" do
       expect(subject.delete(["a", "b", "items"])).to eq("a" => :a, "b" => :b, "items" => {items: "x"})
-    end
-  end
-
-  describe ".reorder" do
-    it "returns reordered array by given options" do
-      result = described_class.reorder(list,
-                      [:a, :b] => {after: :c},
-                      d: {before: :c},
-                      g: :first,
-                      [:e, :f] => {before: :g},
-                      c:  {after: :items},
-                      items: :first
-                     )
-      expect(result).to eq([{:items=>"xx"}, "e", "f", "g", "d", "a", "b", "c"])
     end
   end
 end