+++ templates/releaf/installer.rb
@@ -37,6 +37,7 @@ application "config.i18n.enforce_available_locales = true"
 
 # in "test" env "true" cause to fail on install generators, revert to normall
 gsub_file 'config/environments/test.rb', 'config.cache_classes = false', 'config.cache_classes = true'
+gsub_file 'config/environments/test.rb', 'config.action_controller.allow_forgery_protection = false', 'config.action_controller.allow_forgery_protection = true'
 rake 'db:migrate'
 rake 'db:seed'